Fix Errors
Before, when you first loaded the extension and clicked generate link, it would work but the date and time were missing. Now, the date prop in state is a moment object and set when the component mounts. This way, you can click the link as soon as the app loads. I also made it so if you change the date or time, the link is removed. This is not optimal but solves the problem now.